perf_buffer: use MMap from ring_buf
